Añadido módulo de búsqueda, separando el mismo de la función de gestión de juegos
